rename memory_store to mock_kv
